+++ src/lib/features/export-import-toggles/export-import-service.ts
@@ -10,7 +10,7 @@ import { IFeatureTagStore } from '../../types/stores/feature-tag-store';
 import { ITagTypeStore } from '../../types/stores/tag-type-store';
 import { IStrategy } from '../../types/stores/strategy-store';
 import { IFeatureToggleStore } from '../feature-toggle/types/feature-toggle-store-type';
-import { IFeatureStrategiesStore } from '../feature-toggle/types/feature-strategies-store-type';
+import { IFeatureStrategiesStore } from '../feature-toggle/types/feature-toggle-strategies-store-type';
 import { IFeatureEnvironmentStore } from '../../types/stores/feature-environment-store';
 import { IContextFieldStore, IUnleashStores } from '../../types/stores';
 import { ISegmentStore } from '../../types/stores/segment-store';
