+++ src/lib/routes/admin-api/project/index.ts
@@ -8,7 +8,7 @@ import {
     NONE,
     serializeDates,
 } from '../../../types';
-import ProjectFeaturesController from '../../../features/feature-toggle/feature-controller';
+import ProjectFeaturesController from '../../../features/feature-toggle/feature-toggle-controller';
 import EnvironmentsController from './environments';
 import ProjectHealthReport from './health-report';
 import ProjectService from '../../../services/project-service';
